services/alm_integrator/requirements.txt
google-cloud-pubsub==2.13.0
